Add "rxjava3" to package name
