+++ rxandroid/src/main/java/io/reactivex/rxjava3/android/MainThreadDisposable.java
@@ -11,10 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivex.android;
+package io.reactivex.rxjava3.android;
 
 import android.os.Looper;
-import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers;
 import io.reactivex.rxjava3.disposables.Disposable;
 import java.util.concurrent.atomic.AtomicBoolean;
