+++ rxandroid/src/main/java/io/reactivex/rxjava3/android/schedulers/HandlerScheduler.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivex.android.schedulers;
+package io.reactivex.rxjava3.android.schedulers;
 
 import android.annotation.SuppressLint;
 import android.os.Handler;
