+++ rxandroid/src/test/java/io/reactivex/rxjava3/android/schedulers/HandlerSchedulerTest.java
@@ -11,13 +11,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivex.android.schedulers;
+package io.reactivex.rxjava3.android.schedulers;
 
 import android.os.Handler;
 import android.os.Looper;
 import android.os.Message;
 
-import io.reactivex.android.testutil.CountingRunnable;
+import io.reactivex.rxjava3.android.testutil.CountingRunnable;
+import io.reactivex.rxjava3.android.schedulers.HandlerScheduler;
 import io.reactivex.rxjava3.core.Scheduler;
 import io.reactivex.rxjava3.core.Scheduler.Worker;
 import io.reactivex.rxjava3.disposables.Disposable;
